@@ -39,14 +39,14 @@
 	"dependencies": {
 		"@nextcloud/auth": "^2.5.1",
 		"@nextcloud/axios": "^2.1.0",
-		"@nextcloud/dialogs": "^6.3.0",
-		"@nextcloud/initial-state": "^2.0.0",
+		"@nextcloud/dialogs": "^7.2.0",
+		"@nextcloud/initial-state": "^3.0.0",
 		"@nextcloud/l10n": "^3.2.0",
 		"@nextcloud/moment": "^1.3.4",
-		"@nextcloud/password-confirmation": "^5.3.1",
+		"@nextcloud/password-confirmation": "^6.0.2",
 		"@nextcloud/router": "^3.0.1",
-		"@nextcloud/vue": "^8.26.1",
-		"vue": "^2.7.12",
+		"@nextcloud/vue": "^9.3.1",
+		"vue": "^3.5.26",
 		"vue-html-secure": "^1.0.10",
 		"vue-material-design-icons": "^5.3.1"
 	},
@@ -55,8 +55,8 @@
 		"@nextcloud/browserslist-config": "^3.0.1",
 		"@nextcloud/eslint-config": "^8.4.2",
 		"@nextcloud/stylelint-config": "^3.0.1",
-		"@nextcloud/vite-config": "^1.4.2",
+		"@nextcloud/vite-config": "^2.5.2",
 		"vite-plugin-eslint": "^1.8.1",
-		"vite-plugin-stylelint": "^5.3.1"
+		"vite-plugin-stylelint": "^6.0.2"
 	}
 }
