@@ -11,12 +11,12 @@
 
 document.addEventListener('DOMContentLoaded', () => {
 	OCA.Dashboard.register('zammad_notifications', async (el, { widget }) => {
-		const { default: Vue } = await import('vue')
+		const { createApp } = await import('vue')
 		const { default: Dashboard } = await import('./views/Dashboard.vue')
-		Vue.mixin({ methods: { t, n } })
-		const View = Vue.extend(Dashboard)
-		new View({
-			propsData: { title: widget.title },
-		}).$mount(el)
+		const app = createApp(Dashboard, {
+			title: widget.title,
+		})
+		app.mixin({ methods: { t, n } })
+		app.mount(el)
 	})
 })
