@@ -9,9 +9,9 @@
  * @copyright Julien Veyssier 2020
  */
 
-import Vue from 'vue'
+import { createApp } from 'vue'
 import PersonalSettings from './components/PersonalSettings.vue'
-Vue.mixin({ methods: { t, n } })
 
-const View = Vue.extend(PersonalSettings)
-new View().$mount('#zammad_prefs')
+const app = createApp(PersonalSettings)
+app.mixin({ methods: { t, n } })
+app.mount('#zammad_prefs')
